pyproject.toml
[tool.poetry]
name = "amsr_experiments"
version = "0.1.0"
description = "Research codebase for AMSR+ experiments"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
wandb = "^0.16.0"
optuna = "^3.4.0"
torch = "^2.0.0"
torchvision = "^0.15.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pandas = "^2.1.0"
pyyaml = "^6.0.0"
scikit-learn = "^1.3.0"
scipy = "^1.11.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
